Extractinv and Refactoring the Results implementation, so it's reusable across the Endatix Hub
